+++ Assets/Editor/Tutorials/TutorialCallbacks.cs
@@ -4,19 +4,19 @@ using Unity.Tutorials.Core.Editor;
 
 namespace Unity.Netcode.Samples.BossRoom
 {
-
     /// <summary>
     /// Implement your Tutorial callbacks here.
     /// </summary>
-    [CreateAssetMenu(fileName = DefaultFileName, menuName = "Tutorials/" + DefaultFileName + " Instance")]
+    [CreateAssetMenu(fileName = k_DefaultFileName, menuName = "Tutorials/" + k_DefaultFileName + " Instance")]
     public class TutorialCallbacks : ScriptableObject
     {
-        [SerializeField] SceneAsset m_StartupScene;
+        [SerializeField]
+        SceneAsset m_StartupScene;
 
         /// <summary>
         /// The default file name used to create asset of this class type.
         /// </summary>
-        public const string DefaultFileName = "TutorialCallbacks";
+        const string k_DefaultFileName = "TutorialCallbacks";
 
         /// <summary>
         /// Creates a TutorialCallbacks asset and shows it in the Project window.
@@ -27,7 +27,7 @@ namespace Unity.Netcode.Samples.BossRoom
         /// <returns>The created asset</returns>
         public static ScriptableObject CreateAndShowAsset(string assetPath = null)
         {
-            assetPath = assetPath ?? $"{TutorialEditorUtils.GetActiveFolderPath()}/{DefaultFileName}.asset";
+            assetPath = assetPath ?? $"{TutorialEditorUtils.GetActiveFolderPath()}/{k_DefaultFileName}.asset";
             var asset = CreateInstance<TutorialCallbacks>();
             AssetDatabase.CreateAsset(asset, AssetDatabase.GenerateUniqueAssetPath(assetPath));
             EditorUtility.FocusProjectWindow(); // needed in order to make the selection of newly created asset to really work
@@ -40,35 +40,14 @@ namespace Unity.Netcode.Samples.BossRoom
             TutorialWindow.StartTutorial(tutorial);
         }
 
-        public void FocusGameView()
+        public bool IsConnectedToUgs()
         {
-            /*
-             * note: this solution is a bit weak, but it's the best we can do without accessing internal APIs.
-             * we'll need to check that it works for Unity 6 as well
-             */
-            EditorApplication.ExecuteMenuItem("Window/General/Game");
+            return CloudProjectSettings.projectBound;
         }
 
-        public void FocusSceneView()
+        public void ShowServicesSettings()
         {
-            EditorApplication.ExecuteMenuItem("Window/General/Scene");
-        }
-
-        public bool IsRunningAsHost()
-        {
-            return NetworkManager.Singleton && NetworkManager.Singleton.IsHost;
-        }
-
-        public bool IsRunningAsServerOnly()
-        {
-            return NetworkManager.Singleton && NetworkManager.Singleton.IsServer
-                                            && !NetworkManager.Singleton.IsClient;
-        }
-
-        public bool IsRunningAsClientOnly()
-        {
-            return NetworkManager.Singleton && !NetworkManager.Singleton.IsServer
-                                            && NetworkManager.Singleton.IsClient;
+            SettingsService.OpenProjectSettings("Project/Services");
         }
 
         public void OpenURL(string url)