+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Additional documentation and release notes are available at [Multiplayer Documen
 * Added a welcome dialog to guide users on their first time experience (#888)
   * Added a Table of Contents with helpful resources, that can be accessed from the `Tutorials > Show Tutorials` menu
 * Added the Multiplayer Play Mode package v1.3.1 to the project (#890)
+* An In-Editor Tutorial walking through the process of associating your Unity project with a Unity Cloud ID has been added (#898)
 
 ### Changed
 * Upgraded editor version to 6000.0.25f1 and upgraded Boss Room to Netcode for GameObjects v2.0.0 (#890)